split layout builder to layout builder and doc builder
